Renaming tokenizer_spec to tokens as we bring in more functionality
